+++ ensawards.org/data/ens-best-practices/styles.ts
@@ -1,0 +1,3 @@
+// Separated from the other utils to avoid circular dependency issues.
+export const bestPracticeTechnicalDetailsLinkStyles =
+  "text-black underline underline-offset-[25%] decoration-black/40 hover:decoration-black transition-all duration-200";
